requirements.txt
boto3==1.9.130
line_bot_sdk==1.8.0